Removed project PythonInterface, created NeuralNet.cs
